+++ src/pages/distroSettings/NewDistro/NewDistroButton.test.tsx
