+++ src/pages/distroSettings/NewDistro/newDistroSchema.ts
